\name{is.linnet}
\alias{is.linnet}
\title{Test Whether An Object Is A Linear Network}
\description{
  Checks whether its argument is a linear network
  (object of class \code{"linnet"}).
}
\usage{
is.linnet(x)
}
\arguments{
  \item{x}{Any object.}
}
\details{
  This function tests whether the object \code{x}
  is a linear network (object of class \code{"linnet"}). 
}
\value{
  \code{TRUE} if \code{x} is of class \code{"linnet"},
  otherwise \code{FALSE}.
}
\author{
  \spatstatAuthors.  
}
\keyword{spatial}
\keyword{manip}
