\name{spatstat.geom-internal} 
\title{Internal spatstat.geom functions}
\alias{[.diagramobj}
\alias{[.pp3}
\alias{[.rat}
\alias{[.splitppx}
\alias{[<-.splitppx}


%%%%%%%
\description{
  Internal spatstat.geom functions.
}
\usage{
\method{[}{diagramobj}(x, \dots)
\method{[}{rat}(x, \dots)
\method{[}{splitppx}(x, \dots)
\method{[}{splitppx}(x, \dots) <- value

}
\details{
  These internal \pkg{spatstat.geom} functions are not usually called
  directly by the user. Their names and capabilities may change
  without warning from one version of \pkg{spatstat.geom} to the next.
}
\keyword{internal}
